mcp_server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blender-mcp"
version = "1.2.2"
description = "MCP server bridging AI assistants to Blender"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Advanced Blender MCP Bridge contributors" },
]
keywords = ["blender", "mcp", "model-context-protocol", "ai", "3d"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "mcp>=1.2",
    "pydantic>=2",
    "websockets==12.0",
    "keyring",
    "python-ulid",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-benchmark",
    "pytest-xdist",
    "ruff",
    "mypy",
    "jsonschema",
]

[project.urls]
Homepage = "https://github.com/yesserkira/Advanced-blender-mcp-bridge"
Repository = "https://github.com/yesserkira/Advanced-blender-mcp-bridge"
Documentation = "https://github.com/yesserkira/Advanced-blender-mcp-bridge/tree/main/docs"
Issues = "https://github.com/yesserkira/Advanced-blender-mcp-bridge/issues"

[project.scripts]
blender-mcp = "blender_mcp.server:main"
blender-mcp-install = "blender_mcp.installer:main"

[tool.hatch.build.targets.sdist]
packages = ["blender_mcp"]

[tool.hatch.build.targets.wheel]
packages = ["blender_mcp"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = [".", ".."]
markers = [
    "bench: micro-benchmarks (skipped unless explicitly selected with -m bench)",
]
# Skip benchmarks in the default suite — they're slow and noisy. Run them
# explicitly with: uv run pytest -m bench --benchmark-only
addopts = "-m 'not bench'"

[tool.mypy]
python_version = "3.11"
strict = true

[dependency-groups]
dev = []
